Increasing readability of code
