+++ PlatformFighter/Assets/Animation/RectJab.anim
@@ -76,7 +76,7 @@ AnimationClip:
     floatParameter: 0
     intParameter: 0
     messageOptions: 0
-  - time: 0.16666667
+  - time: 0.083333336
     functionName: attackLag
     data: 
     objectReferenceParameter: {fileID: 0}
