+++ PlatformFighter/Assets/Animation/RectNeutralB.anim
@@ -25,9 +25,9 @@ AnimationClip:
       value: {fileID: 21300000, guid: e00000562a23e6e498abe02973294283, type: 3}
     - time: 0.083333336
       value: {fileID: 21300000, guid: f665819b2e2936b45bc822d2693f6411, type: 3}
-    - time: 0.125
+    - time: 0.20833333
       value: {fileID: 21300000, guid: a1b6965bc81ef9e498653b8126238d9d, type: 3}
-    - time: 0.16666667
+    - time: 0.25
       value: {fileID: 21300000, guid: a537bcd44bd19e24c9bd7b4ad143cb13, type: 3}
     attribute: m_Sprite
     path: 
@@ -58,7 +58,7 @@ AnimationClip:
     m_AdditiveReferencePoseClip: {fileID: 0}
     m_AdditiveReferencePoseTime: 0
     m_StartTime: 0
-    m_StopTime: 0.20833334
+    m_StopTime: 0.29166666
     m_OrientationOffsetY: 0
     m_Level: 0
     m_CycleOffset: 0
@@ -77,4 +77,11 @@ AnimationClip:
   m_EulerEditorCurves: []
   m_HasGenericRootTransform: 0
   m_HasMotionFloatCurves: 0
-  m_Events: []
+  m_Events:
+  - time: 0.125
+    functionName: shootBullet
+    data: 
+    objectReferenceParameter: {fileID: 0}
+    floatParameter: 0
+    intParameter: 0
+    messageOptions: 0
