+++ PlatformFighter/Assets/Scripts/PlayerAttacks.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 02564b3ff62bbe240a6581b87791f286
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
