+++ PlatformFighter/Assets/Scripts/PlayerComputations.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c6e3f657f1194454eb53f535d68e444e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
