+++ PlatformFighter/Assets/Scripts/PlayerHelperFunctions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e808f0ee49778c0499edda8d5cfe0bf0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
